updater-version/configs/requirements.txt
kubernetes
gitPython